satellite collesion/requirements.txt
openai
python-dotenv
pandas
numpy
matplotlib
flask
plotly
sqlite3